Make settings import (even) more secure, based on vulnerability reported by Marcin Węgłowski of afine.pl.

+++ classes/autoptimizeCriticalCSSSettingsAjax.php
@@ -282,14 +282,19 @@ class autoptimizeCriticalCSSSettingsAjax {
 
         // Process an uploaded file with no errors.
         if ( current_user_can( 'manage_options' ) && ! $_FILES['file']['error'] && strpos( $_FILES['file']['name'], '.zip' ) === strlen( $_FILES['file']['name'] ) - 4 ) {
-            // Save file to the cache directory.
-            $zipfile = AO_CCSS_DIR . $_FILES['file']['name'];
+            // create tmp dir with hard guess name in AO_CCSS_DIR.
+            $_secret_dir    = wp_hash( AUTOPTIMIZE_CACHE_URL );
+            $_import_tmp_dir = trailingslashit( AO_CCSS_DIR . $_secret_dir );
+            mkdir( $_import_tmp_dir );
+
+            // Save file to that tmp directory.
+            $zipfile = $_import_tmp_dir . $_FILES['file']['name'];
             move_uploaded_file( $_FILES['file']['tmp_name'], $zipfile );
 
-            // Extract archive.
+            // Extract archive in the tmp directory.
             $zip = new ZipArchive;
             if ( $zip->open( $zipfile ) === true ) {
-                $zip->extractTo( AO_CCSS_DIR );
+                $zip->extractTo( $_import_tmp_dir );
                 $zip->close();
             } else {
                 $error = 'could not extract';
@@ -297,16 +302,28 @@ class autoptimizeCriticalCSSSettingsAjax {
 
             if ( ! $error ) {
                 // only known files allowed, all others are deleted.
-                $_dir_contents_ccss = glob( AO_CCSS_DIR . 'ccss_*.css' );
-                $_dir_known_ok      = array( AO_CCSS_DIR . 'queue.lock', AO_CCSS_DIR . 'queuelog.html', AO_CCSS_DIR . 'index.html', AO_CCSS_DIR . 'settings.json' );
+                $_dir_contents_ccss = glob( $_import_tmp_dir . 'ccss_*.css' );
+                $_dir_known_ok      = array( $_import_tmp_dir . 'settings.json' );
                 $_dir_contents_ok   = array_merge( $_dir_contents_ccss, $_dir_known_ok );
-                $_dir_contents_all  = glob( AO_CCSS_DIR . '*' );
+                $_dir_contents_all  = glob( $_import_tmp_dir . '*' );
                 $_dir_to_be_deleted = array_diff( $_dir_contents_all, $_dir_contents_ok );
                 foreach ( $_dir_to_be_deleted as $_file_to_be_deleted ) {
-                    unlink( $_file_to_be_deleted );
+                    if ( is_dir( $_file_to_be_deleted ) ) {
+                        $this->rrmdir( $_file_to_be_deleted );
+                    } else {
+                        unlink( $_file_to_be_deleted );
+                    }
                 }
 
-                // Archive extraction ok, continue settings importing
+                // and then move known good files to AO_CCSS_DIR.
+                foreach ( $_dir_contents_ok as $ok_file ) {
+                    rename( $ok_file, str_replace( $_secret_dir, '', $ok_file ) );
+                }
+                
+                // and then remove the tmp dir recursively.
+                $this->rrmdir( $_import_tmp_dir );
+
+                // Archive extraction ok, continue importing settings from AO_CCSS_DIR.
                 // Settings file.
                 $importfile = AO_CCSS_DIR . 'settings.json';
 
@@ -360,5 +377,17 @@ class autoptimizeCriticalCSSSettingsAjax {
         } else {
             return true;
         }
+    }
+    
+    public function rrmdir( $path ) {
+        // recursively remove a directory as found on
+        // https://andy-carter.com/blog/recursively-remove-a-directory-in-php.
+        $files = glob($path . '/*');
+        foreach ( $files as $file ) {
+            is_dir( $file ) ? $this->rrmdir( $file ) : unlink( $file );
+        }
+        rmdir( $path );
+
+        return;
     }
 }
